added po filters, in particuler: no repetition of wd POs in guards

+++ org.eventb.core/src/org/eventb/internal/core/pog/ProofObligationGenerator.java
@@ -51,7 +51,8 @@ public abstract class ProofObligationGenerator  implements IAutomaticTool, IExtr
 		
 			monitor.beginTask(Messages.bind(Messages.build_cleaning, file.getName()), 1);
 			
-			file.delete(true, monitor);
+			if (file.exists())
+				file.delete(true, monitor);
 			
 		} finally {
 			monitor.done();
